requirements.txt
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.0
PyJWT==1.7.1
